display-manager-monitor: modify not to use ps/pgrep commands
As of now, enlightenment_mon.sh script uses ps and pgrep commands to get the
process id and the status of enlightenment and it will lookup the directories
under /proc directory. There are lots of files under /proc directory and
actually they are the resources of another process running.

From the perspective of the security modules, accessing resources of another
process will be recognizable as an abnormal behavior. Thus, modifying
enlightenment_mon.sh has been done and there will be no more access for the file
of another process under /proc directory.

Change-Id: I31a130e4df07cee2bdf3b0823d11fbc59a9d69c4
Signed-off-by: Sung-Jin Park <sj76.park@samsung.com>